Added `mix`.
Basically, it behaves just like `alias`, only now you can use the `into`
keyword to specify what class to mix the templates into. It also adds a `set`
keyword, which is just sugar for the optional hash reference of package
variables.

Also updated the documentation to emphasize `mix` and to deprecate `alias` and
`import_templates`.

Moved the private mixin subs `_import` and `_import_code` to the end of the
package with all the other private subs.