Straightend out relations between mixins and delegations. Restored `alias` as a first-class method. It is responsible for delegation. Changed `mix` so that the invocant to its templtes is always the class into which it was mixed. This is more like mixing in. Restored `import_templates` to its previous behavior of using the class into which the templates were imported as the invocant.
In short:

`mix` is the new name for `import_templates` and is responsible for mixins.
`alias` is restored and is responsible for delegation.
`import_templates` is no longer incompatible with previous versions, but is still deprecated.

With this change, I'm pretty happy with things. I'll finish the documentation
next week. But feel free to merge in the meantime.