Avoid depending on the exact wording of Perl error messages
As of Perl 5.21.7, a invocant-less method call like `htm::div {}` when no
htm::div subroutine exists now yields an error message describing the
absence of the invocant, rather than the nonexistence of the subroutine.

This change avoids looking at the wording of Perl's error message, and
instead directly verifies that no htm::div subroutine is created by such
code (which is what I think the test is trying to ensure).